refactor(asr): share Apple audio model runners
